!refactor(mock-server): determine admin area from port
!BREAKING CHANGE: the admin area used to be calculated via the 'mock-server' path segment
