+++ src/NinjaTools.FluentMockServer.API/Administration/AdminOptions.cs
@@ -1,0 +1,7 @@
+namespace NinjaTools.FluentMockServer.API.Administration
+{
+    public class AdminOptions
+    {
+        public int Port { get; set; }
+    }
+}
