+++ src/NinjaTools.FluentMockServer.API/DependencyInjection/MockServerBuilderExtensions.cs
@@ -1,0 +1,17 @@
+using JetBrains.Annotations;
+using Microsoft.Extensions.DependencyInjection;
+using NinjaTools.FluentMockServer.API.Administration;
+
+namespace NinjaTools.FluentMockServer.API.DependencyInjection
+{
+    public static class MockServerBuilderExtensions
+    {
+        [NotNull]
+        public static IMockServerBuilder AddAdminPath([NotNull] this IMockServerBuilder builder, int port = 2081)
+        {
+            builder.Services.Configure<AdminOptions>(opt => { opt.Port = port; });
+
+            return builder;
+        }
+    }
+}
